more start improvements
