+++ src/util/icons.tsx
@@ -35,6 +35,7 @@ export const iconCog = "M12 15C10.343 15 9 13.657 9 12 9 10.343 10.343 9 12 9 13
 export const iconCombined = "M21 6.949 8 6.949 8 6.051 21 6.051ZM21 11.949 8 11.949 8 11.051 21 11.051ZM21 16.949 8 16.949 8 16.051 21 16.051ZM7.5 3 7.5 4 6.7 4C6.132 4 5.736 4 5.427 4.026 5.125 4.05 4.951 4.096 4.819 4.163 4.537 4.307 4.307 4.537 4.163 4.819 4.096 4.951 4.05 5.125 4.026 5.427 4 5.736 4 6.132 4 6.7L4 16.3C4 16.868 4 17.264 4.026 17.573 4.05 17.875 4.096 18.049 4.163 18.181 4.307 18.463 4.537 18.693 4.819 18.837 4.951 18.904 5.125 18.95 5.427 18.974 5.736 19 6.132 19 6.7 19L7.5 19 7.5 20 6.679 20C6.137 20 5.7 20 5.346 19.971 4.981 19.941 4.661 19.878 4.365 19.728 3.895 19.488 3.512 19.105 3.272 18.635 3.122 18.339 3.059 18.019 3.029 17.654 3 17.301 3 16.864 3 16.322L3 6.678C3 6.136 3 5.7 3.029 5.346 3.059 4.981 3.122 4.661 3.272 4.365 3.512 3.895 3.895 3.512 4.365 3.272 4.661 3.122 4.981 3.059 5.346 3.029 5.7 3 6.136 3 6.678 3L7.5 3Z";
 export const iconCommand = "M3,6.5 C3,4.567 4.567,3 6.5,3 C8.433,3 10,4.567 10,6.5 L10,9 L14,9 L14,6.5 C14,4.567 15.567,3 17.5,3 C19.433,3 21,4.567 21,6.5 C21,8.433 19.433,10 17.5,10 L15,10 L15,14 L17.5,14 C19.433,14 21,15.567 21,17.5 C21,19.433 19.433,21 17.5,21 C15.567,21 14,19.433 14,17.5 L14,15 L10,15 L10,17.5 C10,19.433 8.433,21 6.5,21 C4.567,21 3,19.433 3,17.5 C3,15.567 4.567,14 6.5,14 L9,14 L9,10 L6.5,10 C4.567,10 3,8.433 3,6.5 Z M9,6.5 C9,5.119 7.881,4 6.5,4 C5.119,4 4,5.119 4,6.5 C4,7.881 5.119,9 6.5,9 L9,9 Z M15,6.5 L15,9 L17.5,9 C18.881,9 20,7.881 20,6.5 C20,5.119 18.881,4 17.5,4 C16.119,4 15,5.119 15,6.5 Z M9,15 L6.5,15 C5.119,15 4,16.119 4,17.5 C4,18.881 5.119,20 6.5,20 C7.881,20 9,18.881 9,17.5 Z M15,17.5 C15,18.881 16.119,20 17.5,20 C18.881,20 20,18.881 20,17.5 C20,16.119 18.881,15 17.5,15 L15,15 Z M10,10 L10,14 L14,14 L14,10 Z";
 export const iconComment = "M2.5,5.5 C2.5,3.843 3.843,2.5 5.5,2.5 L18.5,2.5 C20.157,2.5 21.5,3.843 21.5,5.5 L21.5,15 C21.5,16.657 20.157,18 18.5,18 L12.112,18 L8.517,21.318 C7.557,22.205 6,21.524 6,20.216 L6,18 L5.5,18 C3.843,18 2.5,16.657 2.5,15 Z M5.5,3.5 C4.395,3.5 3.5,4.395 3.5,5.5 L3.5,15 C3.5,16.105 4.395,17 5.5,17 L7,17 L7,20.216 C7,20.652 7.519,20.879 7.839,20.583 L11.721,17 L18.5,17 C19.605,17 20.5,16.105 20.5,15 L20.5,5.5 C20.5,4.395 19.605,3.5 18.5,3.5 Z";
+export const iconCommunity = "M8.021,9.02 C8.02,8.864 8.02,8.698 8.02,8.519 L8.02,5.479 C8.02,4.972 8.02,4.562 8.047,4.229 C8.075,3.887 8.134,3.585 8.276,3.306 C8.502,2.862 8.862,2.502 9.306,2.276 C9.585,2.134 9.887,2.075 10.229,2.047 C10.562,2.02 10.972,2.02 11.479,2.02 L18.52,2.02 C19.027,2.02 19.437,2.02 19.77,2.047 C20.112,2.075 20.414,2.134 20.693,2.276 C21.137,2.502 21.497,2.862 21.723,3.306 C21.865,3.585 21.924,3.887 21.952,4.229 C21.979,4.562 21.979,4.972 21.979,5.48 L21.979,8.519 C21.979,9.027 21.979,9.437 21.952,9.77 C21.924,10.112 21.865,10.414 21.723,10.693 C21.497,11.137 21.137,11.497 20.693,11.723 C20.382,11.882 20.041,11.937 19.646,11.961 C19.559,11.966 19.516,12.031 19.516,12.08 L19.522,13.584 C19.525,14.213 18.772,14.538 18.316,14.105 L16.108,12.009 C16.089,11.99 16.063,11.979 16.035,11.979 L14.979,11.979 C14.98,12.135 14.98,12.301 14.98,12.48 L14.98,15.519 C14.98,16.027 14.98,16.437 14.953,16.77 C14.925,17.112 14.866,17.414 14.723,17.693 C14.498,18.137 14.137,18.497 13.694,18.723 C13.414,18.865 13.113,18.924 12.77,18.952 C12.438,18.979 12.028,18.979 11.52,18.979 L6.642,18.979 C6.615,18.979 6.589,18.99 6.569,19.009 L4.184,21.274 C3.727,21.707 2.975,21.382 2.977,20.753 L2.984,19.046 C2.984,19.023 2.976,18.997 2.95,18.969 C2.921,18.938 2.873,18.908 2.812,18.895 C2.633,18.858 2.466,18.804 2.306,18.723 C1.863,18.497 1.503,18.137 1.277,17.693 C1.134,17.414 1.075,17.112 1.047,16.77 C1.02,16.437 1.02,16.027 1.02,15.519 L1.02,12.48 C1.02,11.972 1.02,11.562 1.047,11.229 C1.075,10.887 1.134,10.585 1.277,10.306 C1.503,9.862 1.863,9.502 2.306,9.276 C2.586,9.134 2.887,9.075 3.23,9.047 C3.562,9.02 3.972,9.02 4.48,9.02 Z M2.742,10.132 C2.479,10.265 2.266,10.479 2.132,10.741 C2.07,10.863 2.027,11.024 2.004,11.308 C1.98,11.596 1.98,11.967 1.98,12.5 L1.98,15.499 C1.98,16.032 1.98,16.403 2.004,16.691 C2.027,16.975 2.07,17.136 2.132,17.258 C2.266,17.52 2.479,17.734 2.742,17.867 C2.812,17.903 2.895,17.932 3.006,17.955 C3.482,18.053 3.946,18.455 3.944,19.05 L3.939,20.182 L5.908,18.313 C6.106,18.124 6.369,18.02 6.642,18.02 L11.499,18.02 C12.032,18.02 12.403,18.019 12.692,17.996 C12.975,17.972 13.136,17.929 13.258,17.867 C13.521,17.734 13.734,17.52 13.868,17.258 C13.93,17.136 13.973,16.975 13.996,16.691 C14.02,16.403 14.02,16.032 14.02,15.499 L14.02,12.5 C14.02,11.967 14.02,11.596 13.996,11.308 C13.973,11.024 13.93,10.863 13.868,10.741 C13.734,10.479 13.521,10.265 13.258,10.132 C13.136,10.07 12.975,10.027 12.692,10.003 C12.403,9.98 12.032,9.98 11.499,9.98 L4.501,9.98 C3.968,9.98 3.597,9.98 3.308,10.003 C3.025,10.027 2.864,10.07 2.742,10.132 Z M19.691,3.003 C19.403,2.98 19.032,2.98 18.499,2.98 L11.5,2.98 C10.967,2.98 10.596,2.98 10.308,3.003 C10.024,3.027 9.863,3.07 9.741,3.132 C9.479,3.265 9.265,3.479 9.132,3.741 C9.07,3.863 9.027,4.024 9.003,4.308 C8.98,4.596 8.98,4.967 8.98,5.5 L8.98,8.499 C8.98,8.693 8.98,8.865 8.981,9.02 L11.52,9.02 C12.028,9.02 12.438,9.02 12.77,9.047 C13.113,9.075 13.414,9.134 13.694,9.276 C14.137,9.502 14.498,9.862 14.723,10.306 C14.836,10.527 14.896,10.762 14.931,11.019 L16.035,11.019 C16.309,11.019 16.572,11.124 16.77,11.313 L18.56,13.013 L18.556,12.084 C18.554,11.487 19.032,11.036 19.589,11.003 C19.936,10.982 20.121,10.937 20.258,10.867 C20.52,10.734 20.734,10.52 20.867,10.258 C20.929,10.136 20.972,9.975 20.996,9.692 C21.019,9.403 21.02,9.032 21.02,8.499 L21.02,5.5 C21.02,4.967 21.019,4.596 20.996,4.308 C20.972,4.024 20.929,3.863 20.867,3.741 C20.734,3.479 20.52,3.265 20.258,3.132 C20.136,3.07 19.975,3.027 19.691,3.003 Z";
 export const iconConsole = "M5.239 6.317 5.874 5.683 10.474 10.283C11.391 11.2 11.391 12.687 10.474 13.604L5.874 18.204 5.239 17.569 9.839 12.969C10.406 12.403 10.406 11.484 9.839 10.917ZM20 19 10.5 19 10.5 18 20 18 20 19Z";
 export const iconCopy = "M8.478 2 12.522 2C13.203 2 13.745 2 14.183 2.036 14.63 2.072 15.012 2.149 15.362 2.327 15.927 2.615 16.385 3.074 16.673 3.638 16.873 4.03 16.947 4.468 16.977 5L15.975 5C15.946 4.556 15.886 4.296 15.782 4.092 15.59 3.716 15.284 3.41 14.908 3.218 14.723 3.124 14.487 3.064 14.101 3.032 13.71 3 13.208 3 12.5 3L8.5 3C7.792 3 7.291 3 6.899 3.032 6.513 3.064 6.277 3.124 6.092 3.218 5.716 3.41 5.41 3.716 5.218 4.092 5.124 4.277 5.064 4.513 5.032 4.899 5 5.291 5 5.792 5 6.5L5 14.5C5 15.208 5 15.71 5.032 16.101 5.064 16.487 5.124 16.723 5.218 16.908 5.41 17.284 5.716 17.59 6.092 17.782 6.296 17.886 6.556 17.946 7 17.975L7 18.977C6.468 18.947 6.03 18.873 5.638 18.673 5.074 18.385 4.615 17.927 4.327 17.362 4.149 17.012 4.072 16.631 4.036 16.183 4 15.745 4 15.203 4 14.522L4 6.478C4 5.797 4 5.255 4.036 4.817 4.072 4.369 4.149 3.988 4.327 3.638 4.615 3.074 5.074 2.615 5.638 2.327 5.988 2.149 6.369 2.072 6.817 2.036 7.255 2 7.797 2 8.478 2ZM8 10C8 8.6 8 7.9 8.272 7.365 8.512 6.895 8.895 6.512 9.365 6.272 9.9 6 10.6 6 12 6L16 6C17.4 6 18.1 6 18.635 6.272 19.105 6.512 19.488 6.895 19.728 7.365 20 7.9 20 8.6 20 10L20 18C20 19.4 20 20.1 19.728 20.635 19.488 21.105 19.105 21.488 18.635 21.728 18.1 22 17.4 22 16 22L12 22C10.6 22 9.9 22 9.365 21.728 8.895 21.488 8.512 21.105 8.272 20.635 8 20.1 8 19.4 8 18ZM12 7C11.283 7 10.806 7.001 10.439 7.031 10.084 7.06 9.922 7.111 9.819 7.163 9.537 7.307 9.307 7.537 9.163 7.819 9.111 7.922 9.06 8.084 9.031 8.439 9.001 8.806 9 9.283 9 10L9 18C9 18.717 9.001 19.194 9.031 19.561 9.06 19.916 9.111 20.078 9.163 20.181 9.307 20.463 9.537 20.693 9.819 20.837 9.922 20.889 10.084 20.94 10.439 20.969 10.806 20.999 11.283 21 12 21L16 21C16.717 21 17.194 20.999 17.561 20.969 17.916 20.94 18.078 20.889 18.181 20.837 18.463 20.693 18.693 20.463 18.837 20.181 18.889 20.078 18.94 19.916 18.969 19.561 18.999 19.194 19 18.717 19 18L19 10C19 9.283 18.999 8.806 18.969 8.439 18.94 8.084 18.889 7.922 18.837 7.819 18.693 7.537 18.463 7.307 18.181 7.163 18.078 7.111 17.916 7.06 17.561 7.031 17.194 7.001 16.717 7 16 7Z";
 export const iconCreditCard = "M5.679,5 L18.322,5 C18.864,5 19.301,5 19.654,5.029 C20.019,5.059 20.339,5.122 20.635,5.272 C21.105,5.512 21.488,5.895 21.728,6.365 C21.878,6.661 21.941,6.981 21.971,7.346 C21.987,7.538 21.994,7.755 21.997,8 L22,8 L22,16.321 C22,16.863 22,17.3 21.971,17.654 C21.941,18.019 21.878,18.339 21.728,18.635 C21.488,19.105 21.105,19.488 20.635,19.728 C20.339,19.878 20.019,19.941 19.654,19.971 C19.3,20 18.863,20 18.322,20 L5.679,20 C5.137,20 4.7,20 4.346,19.971 C3.981,19.941 3.661,19.878 3.365,19.728 C2.895,19.488 2.512,19.105 2.272,18.635 C2.122,18.339 2.059,18.019 2.029,17.654 C2,17.301 2,16.864 2,16.322 L2,8 L2.003,8 C2.006,7.755 2.013,7.538 2.029,7.346 C2.059,6.981 2.122,6.661 2.272,6.365 C2.512,5.895 2.895,5.512 3.365,5.272 C3.661,5.122 3.981,5.059 4.346,5.029 C4.7,5 5.136,5 5.678,5 L5.679,5 Z M3,13 L3,16.3 C3,16.868 3,17.264 3.026,17.573 C3.05,17.875 3.096,18.049 3.163,18.181 C3.307,18.463 3.537,18.693 3.819,18.837 C3.951,18.904 4.125,18.95 4.427,18.974 C4.736,19 5.132,19 5.7,19 L18.3,19 C18.868,19 19.264,19 19.573,18.974 C19.875,18.95 20.049,18.904 20.181,18.837 C20.463,18.693 20.693,18.463 20.837,18.181 C20.904,18.049 20.95,17.875 20.974,17.573 C21,17.264 21,16.868 21,16.3 L21,13 Z M21,9 L3,9 L3,12 L21,12 Z M20.997,8 C20.994,7.776 20.988,7.59 20.974,7.427 C20.95,7.125 20.904,6.951 20.837,6.819 C20.693,6.537 20.463,6.307 20.181,6.163 C20.049,6.096 19.875,6.05 19.573,6.026 C19.264,6 18.868,6 18.3,6 L5.7,6 C5.132,6 4.736,6 4.427,6.026 C4.125,6.05 3.951,6.096 3.819,6.163 C3.537,6.307 3.307,6.537 3.163,6.819 C3.096,6.951 3.05,7.125 3.026,7.427 C3.012,7.59 3.006,7.776 3.003,8 Z";
@@ -44,7 +45,7 @@ export const iconDatabase = "M22,16 C22,17.772 20.763,19.282 18.968,20.322 C17.1
 export const iconDatabaseSecure = "M15,18.4 C15,17.56 15,17.14 15.163,16.819 C15.307,16.537 15.537,16.307 15.819,16.163 C15.876,16.135 15.935,16.111 16,16.091 L16,14.5 C16,12.843 17.343,11.5 19,11.5 C20.657,11.5 22,12.843 22,14.5 L22,16.091 C22.065,16.111 22.124,16.135 22.181,16.163 C22.463,16.307 22.693,16.537 22.837,16.819 C23,17.14 23,17.56 23,18.4 L23,20.6 C23,21.44 23,21.86 22.837,22.181 C22.693,22.463 22.463,22.693 22.181,22.837 C21.86,23 21.44,23 20.6,23 L17.4,23 C16.56,23 16.14,23 15.819,22.837 C15.537,22.693 15.307,22.463 15.163,22.181 C15,21.86 15,21.44 15,20.6 Z M21,7 C21,5.735 20.11,4.495 18.467,3.544 C16.837,2.6 14.551,2 12,2 C9.449,2 7.163,2.6 5.533,3.544 C3.89,4.495 3,5.735 3,7 C3,8.265 3.89,9.505 5.533,10.456 C7.163,11.4 9.449,12 12,12 C12.338,12 12.672,11.99 13,11.969 L13,12.971 C12.671,12.99 12.337,13 12,13 C9.305,13 6.84,12.369 5.032,11.322 C4.219,10.851 3.521,10.284 3,9.64 L3,11.5 C3,12.765 3.89,14.005 5.533,14.956 C7.163,15.9 9.449,16.5 12,16.5 C12.338,16.5 12.672,16.49 13,16.469 L13,17.471 C12.671,17.49 12.337,17.5 12,17.5 C9.305,17.5 6.84,16.869 5.032,15.822 C4.219,15.351 3.521,14.784 3,14.14 L3,16 C3,17.265 3.89,18.505 5.533,19.456 C7.163,20.4 9.449,21 12,21 C12.338,21 12.672,20.99 13,20.969 L13,21.971 C12.671,21.99 12.337,22 12,22 C9.305,22 6.84,21.369 5.032,20.322 C3.237,19.282 2,17.772 2,16 L2,7 C2,5.228 3.237,3.717 5.032,2.678 C6.84,1.631 9.305,1 12,1 C14.695,1 17.16,1.631 18.968,2.678 C20.763,3.717 22,5.228 22,7 C22,7.928 21.657,8.792 21.073,9.548 L20.281,8.937 C20.755,8.324 21,7.67 21,7 Z M17.4,17 C16.963,17 16.696,17.001 16.496,17.017 C16.403,17.025 16.347,17.034 16.312,17.042 C16.296,17.046 16.286,17.049 16.281,17.051 C16.276,17.053 16.273,17.054 16.273,17.055 C16.179,17.102 16.102,17.179 16.055,17.273 L16.054,17.275 L16.051,17.281 C16.049,17.286 16.046,17.296 16.042,17.312 C16.034,17.347 16.025,17.403 16.017,17.496 C16.001,17.696 16,17.963 16,18.4 L16,20.6 C16,21.037 16.001,21.304 16.017,21.504 C16.025,21.597 16.034,21.653 16.042,21.688 C16.046,21.704 16.049,21.714 16.051,21.719 C16.053,21.724 16.054,21.727 16.055,21.727 C16.102,21.821 16.179,21.898 16.273,21.945 C16.273,21.946 16.276,21.947 16.281,21.949 C16.286,21.951 16.296,21.954 16.312,21.958 C16.347,21.966 16.403,21.975 16.496,21.983 C16.696,21.999 16.963,22 17.4,22 L20.6,22 C21.037,22 21.304,21.999 21.504,21.983 C21.597,21.975 21.653,21.966 21.688,21.958 C21.704,21.954 21.714,21.951 21.719,21.949 C21.724,21.947 21.727,21.946 21.727,21.945 C21.821,21.898 21.898,21.821 21.945,21.727 C21.946,21.727 21.947,21.724 21.949,21.719 C21.951,21.714 21.954,21.704 21.958,21.688 C21.966,21.653 21.975,21.597 21.983,21.504 C21.999,21.304 22,21.037 22,20.6 L22,18.4 C22,17.963 21.999,17.696 21.983,17.496 C21.975,17.403 21.966,17.347 21.958,17.312 C21.954,17.296 21.951,17.286 21.949,17.281 C21.947,17.276 21.946,17.273 21.945,17.273 C21.898,17.179 21.821,17.102 21.727,17.055 L21.725,17.054 L21.719,17.051 C21.714,17.049 21.704,17.046 21.688,17.042 C21.653,17.034 21.597,17.025 21.504,17.017 C21.304,17.001 21.037,17 20.6,17 Z M21,16.001 L21,14.5 C21,13.395 20.105,12.5 19,12.5 C17.895,12.5 17,13.395 17,14.5 L17,16.001 C17.122,16 17.255,16 17.4,16 L20.6,16 C20.745,16 20.878,16 21,16.001 Z";
 export const iconDelete = "M12.355 12 9.678 9.323 10.323 8.678 13 11.355 15.677 8.678 16.322 9.323 13.645 12 16.323 14.678 15.678 15.323 13 12.645 10.322 15.323 9.677 14.678 12.355 12ZM5.759 4.602C6.016 4.381 6.315 4.216 6.638 4.114 7.003 4 7.408 4 8.217 4L18 4C19.4 4 20.1 4 20.635 4.272 21.105 4.512 21.488 4.895 21.727 5.365 22 5.9 22 6.6 22 8L22 16C22 17.4 22 18.1 21.727 18.635 21.488 19.105 21.105 19.488 20.635 19.728 20.1 20 19.4 20 18 20L8.217 20C7.408 20 7.003 20 6.638 19.886 6.315 19.784 6.016 19.619 5.759 19.398 5.468 19.149 5.254 18.806 4.825 18.12L2.06 13.696C1.675 13.081 1.483 12.773 1.408 12.443 1.342 12.151 1.342 11.849 1.408 11.557 1.483 11.227 1.675 10.919 2.06 10.304L4.825 5.88C5.254 5.194 5.468 4.851 5.759 4.602ZM18 5 8.217 5C7.343 5 7.121 5.011 6.937 5.069 6.743 5.129 6.564 5.229 6.409 5.361 6.263 5.487 6.136 5.669 5.673 6.41L2.908 10.834C2.484 11.512 2.412 11.653 2.383 11.778 2.35 11.924 2.35 12.076 2.383 12.222 2.412 12.347 2.484 12.488 2.908 13.166L5.673 17.59C6.136 18.331 6.263 18.513 6.409 18.639 6.564 18.771 6.743 18.871 6.937 18.931 7.121 18.989 7.343 19 8.217 19L18 19C18.716 19 19.194 18.999 19.56 18.969 19.916 18.94 20.078 18.889 20.181 18.837 20.463 18.693 20.693 18.463 20.836 18.181 20.889 18.078 20.94 17.916 20.969 17.561 20.999 17.194 21 16.717 21 16L21 8C21 7.283 20.999 6.806 20.969 6.439 20.94 6.084 20.889 5.922 20.836 5.819 20.693 5.537 20.463 5.307 20.181 5.163 20.078 5.111 19.916 5.06 19.56 5.031 19.194 5.001 18.716 5 18 5Z";
 export const iconDesigner = "M19.429 2C19.778 2 20.07 2 20.308 2.019 20.557 2.04 20.791 2.084 21.013 2.197 21.353 2.37 21.63 2.647 21.803 2.987 21.916 3.209 21.96 3.443 21.98 3.692 22 3.93 22 4.222 22 4.571L22 6.429C22 6.778 22 7.07 21.98 7.308 21.96 7.557 21.916 7.791 21.803 8.013 21.63 8.353 21.353 8.63 21.013 8.803 20.791 8.916 20.557 8.96 20.308 8.981 20.07 9 19.778 9 19.429 9L19 9 19 15 19.429 15C19.778 15 20.07 15 20.308 15.019 20.557 15.04 20.791 15.084 21.013 15.197 21.353 15.37 21.63 15.647 21.803 15.987 21.916 16.209 21.96 16.443 21.98 16.692 22 16.93 22 17.222 22 17.571L22 19.429C22 19.778 22 20.07 21.98 20.308 21.96 20.557 21.916 20.791 21.803 21.013 21.63 21.353 21.353 21.63 21.013 21.803 20.791 21.916 20.557 21.96 20.308 21.98 20.07 22 19.778 22 19.429 22L17.571 22C17.222 22 16.93 22 16.692 21.98 16.443 21.96 16.209 21.916 15.987 21.803 15.647 21.63 15.37 21.353 15.197 21.013 15.084 20.791 15.04 20.557 15.019 20.308 15 20.07 15 19.778 15 19.429L15 19 9 19 9 19.429C9 19.778 9 20.07 8.981 20.308 8.96 20.557 8.916 20.791 8.803 21.013 8.63 21.353 8.353 21.63 8.013 21.803 7.791 21.916 7.557 21.96 7.308 21.98 7.07 22 6.778 22 6.429 22L4.571 22C4.222 22 3.93 22 3.692 21.98 3.443 21.96 3.209 21.916 2.987 21.803 2.647 21.63 2.37 21.353 2.197 21.013 2.084 20.791 2.04 20.557 2.019 20.308 2 20.07 2 19.778 2 19.429L2 17.571C2 17.222 2 16.93 2.019 16.692 2.04 16.443 2.084 16.209 2.197 15.987 2.37 15.647 2.647 15.37 2.987 15.197 3.209 15.084 3.443 15.04 3.692 15.019 3.93 15 4.222 15 4.571 15L5 15 5 9 4.571 9C4.222 9 3.93 9 3.692 8.981 3.443 8.96 3.209 8.916 2.987 8.803 2.647 8.63 2.37 8.353 2.197 8.013 2.084 7.791 2.04 7.557 2.019 7.308 2 7.07 2 6.778 2 6.429L2 4.571C2 4.222 2 3.93 2.019 3.692 2.04 3.443 2.084 3.209 2.197 2.987 2.37 2.647 2.647 2.37 2.987 2.197 3.209 2.084 3.443 2.04 3.692 2.019 3.93 2 4.222 2 4.571 2L6.429 2C6.778 2 7.07 2 7.308 2.019 7.557 2.04 7.791 2.084 8.013 2.197 8.353 2.37 8.63 2.647 8.803 2.987 8.916 3.209 8.96 3.443 8.981 3.692 9 3.93 9 4.222 9 4.571L9 5 15 5 15 4.571C15 4.222 15 3.93 15.019 3.692 15.04 3.443 15.084 3.209 15.197 2.987 15.37 2.647 15.647 2.37 15.987 2.197 16.209 2.084 16.443 2.04 16.692 2.019 16.93 2 17.222 2 17.571 2ZM18 9 17.571 9C17.222 9 16.93 9 16.692 8.981 16.443 8.96 16.209 8.916 15.987 8.803 15.647 8.63 15.37 8.353 15.197 8.013 15.084 7.791 15.04 7.557 15.019 7.308 15 7.07 15 6.778 15 6.429L15 6 9 6 9 6.429C9 6.778 9 7.07 8.981 7.308 8.96 7.557 8.916 7.791 8.803 8.013 8.63 8.353 8.353 8.63 8.013 8.803 7.791 8.916 7.557 8.96 7.308 8.981 7.07 9 6.778 9 6.429 9L6 9 6 15 6.429 15C6.778 15 7.07 15 7.308 15.019 7.557 15.04 7.791 15.084 8.013 15.197 8.353 15.37 8.63 15.647 8.803 15.987 8.916 16.209 8.96 16.443 8.981 16.692 9 16.93 9 17.222 9 17.571L9 18 15 18 15 17.571C15 17.222 15 16.93 15.019 16.692 15.04 16.443 15.084 16.209 15.197 15.987 15.37 15.647 15.647 15.37 15.987 15.197 16.209 15.084 16.443 15.04 16.692 15.019 16.93 15 17.222 15 17.571 15L18 15ZM16.441 16.088C16.289 16.165 16.165 16.289 16.088 16.441 16.059 16.498 16.031 16.586 16.016 16.773 16 16.966 16 17.217 16 17.591L16 19.409C16 19.783 16 20.034 16.016 20.227 16.031 20.414 16.059 20.502 16.088 20.559 16.165 20.711 16.289 20.835 16.441 20.912 16.498 20.941 16.586 20.969 16.773 20.984 16.966 21 17.217 21 17.591 21L19.409 21C19.783 21 20.034 21 20.227 20.984 20.414 20.969 20.502 20.941 20.559 20.912 20.711 20.835 20.835 20.711 20.912 20.559 20.941 20.502 20.969 20.414 20.984 20.227 21 20.034 21 19.783 21 19.409L21 17.591C21 17.217 21 16.966 20.984 16.773 20.969 16.586 20.941 16.498 20.912 16.441 20.835 16.289 20.711 16.165 20.559 16.088 20.502 16.059 20.414 16.031 20.227 16.016 20.034 16 19.783 16 19.409 16L17.591 16C17.217 16 16.966 16 16.773 16.016 16.586 16.031 16.498 16.059 16.441 16.088ZM3.773 16.016C3.586 16.031 3.498 16.059 3.441 16.088 3.289 16.165 3.165 16.289 3.088 16.441 3.059 16.498 3.031 16.586 3.016 16.773 3 16.966 3 17.217 3 17.591L3 19.409C3 19.783 3 20.034 3.016 20.227 3.031 20.414 3.059 20.502 3.088 20.559 3.165 20.711 3.289 20.835 3.441 20.912 3.498 20.941 3.586 20.969 3.773 20.984 3.966 21 4.217 21 4.591 21L6.409 21C6.783 21 7.034 21 7.227 20.984 7.414 20.969 7.502 20.941 7.559 20.912 7.711 20.835 7.835 20.711 7.912 20.559 7.941 20.502 7.969 20.414 7.984 20.227 8 20.034 8 19.783 8 19.409L8 17.591C8 17.217 8 16.966 7.984 16.773 7.969 16.586 7.941 16.498 7.912 16.441 7.835 16.289 7.711 16.165 7.559 16.088 7.502 16.059 7.414 16.031 7.227 16.016 7.034 16 6.783 16 6.409 16L4.591 16C4.217 16 3.966 16 3.773 16.016ZM16.441 3.088C16.289 3.165 16.165 3.289 16.088 3.441 16.059 3.498 16.031 3.586 16.016 3.773 16 3.966 16 4.217 16 4.591L16 6.409C16 6.783 16 7.034 16.016 7.227 16.031 7.414 16.059 7.502 16.088 7.559 16.165 7.711 16.289 7.835 16.441 7.912 16.498 7.941 16.586 7.969 16.773 7.984 16.966 8 17.217 8 17.591 8L19.409 8C19.783 8 20.034 8 20.227 7.984 20.414 7.969 20.502 7.941 20.559 7.912 20.711 7.835 20.835 7.711 20.912 7.559 20.941 7.502 20.969 7.414 20.984 7.227 21 7.034 21 6.783 21 6.409L21 4.591C21 4.217 21 3.966 20.984 3.773 20.969 3.586 20.941 3.498 20.912 3.441 20.835 3.289 20.711 3.165 20.559 3.088 20.502 3.059 20.414 3.031 20.227 3.016 20.034 3 19.783 3 19.409 3L17.591 3C17.217 3 16.966 3 16.773 3.016 16.586 3.031 16.498 3.059 16.441 3.088ZM3.773 3.016C3.586 3.031 3.498 3.059 3.441 3.088 3.289 3.165 3.165 3.289 3.088 3.441 3.059 3.498 3.031 3.586 3.016 3.773 3 3.966 3 4.217 3 4.591L3 6.409C3 6.783 3 7.034 3.016 7.227 3.031 7.414 3.059 7.502 3.088 7.559 3.165 7.711 3.289 7.835 3.441 7.912 3.498 7.941 3.586 7.969 3.773 7.984 3.966 8 4.217 8 4.591 8L6.409 8C6.783 8 7.034 8 7.227 7.984 7.414 7.969 7.502 7.941 7.559 7.912 7.711 7.835 7.835 7.711 7.912 7.559 7.941 7.502 7.969 7.414 7.984 7.227 8 7.034 8 6.783 8 6.409L8 4.591C8 4.217 8 3.966 7.984 3.773 7.969 3.586 7.941 3.498 7.912 3.441 7.835 3.289 7.711 3.165 7.559 3.088 7.502 3.059 7.414 3.031 7.227 3.016 7.034 3 6.783 3 6.409 3L4.591 3C4.217 3 3.966 3 3.773 3.016ZM19.429 1.75Q20.08 1.75 20.328 1.77 20.803 1.809 21.127 1.974 21.722 2.278 22.026 2.873 22.191 3.197 22.23 3.672 22.25 3.92 22.25 4.571L22.25 6.429Q22.25 7.08 22.23 7.328 22.191 7.803 22.026 8.127 21.722 8.722 21.127 9.026 20.803 9.191 20.328 9.23 20.08 9.25 19.429 9.25L19.25 9.25 19.25 14.75 19.429 14.75Q20.08 14.75 20.328 14.77 20.803 14.809 21.127 14.974 21.722 15.278 22.026 15.873 22.191 16.197 22.23 16.672 22.25 16.92 22.25 17.571L22.25 19.429Q22.25 20.08 22.23 20.328 22.191 20.803 22.026 21.127 21.722 21.722 21.127 22.026 20.803 22.191 20.328 22.23 20.08 22.25 19.429 22.25L17.571 22.25Q16.92 22.25 16.672 22.23 16.197 22.191 15.873 22.026 15.278 21.722 14.974 21.127 14.809 20.803 14.77 20.328 14.75 20.08 14.75 19.429L14.75 19.25 9.25 19.25 9.25 19.429Q9.25 20.08 9.23 20.328 9.191 20.803 9.026 21.127 8.722 21.722 8.127 22.026 7.803 22.191 7.328 22.23 7.08 22.25 6.429 22.25L4.571 22.25Q3.92 22.25 3.672 22.23 3.197 22.191 2.873 22.026 2.278 21.722 1.974 21.127 1.809 20.803 1.77 20.328 1.75 20.08 1.75 19.429L1.75 17.571Q1.75 16.92 1.77 16.672 1.809 16.197 1.974 15.873 2.278 15.278 2.873 14.974 3.197 14.809 3.672 14.77 3.92 14.75 4.571 14.75L4.75 14.75 4.75 9.25 4.571 9.25Q3.92 9.25 3.672 9.23 3.197 9.191 2.873 9.026 2.278 8.722 1.974 8.127 1.809 7.803 1.77 7.328 1.75 7.08 1.75 6.429L1.75 4.571Q1.75 3.92 1.77 3.672 1.809 3.197 1.974 2.873 2.278 2.278 2.873 1.974 3.197 1.809 3.672 1.77 3.92 1.75 4.571 1.75L6.429 1.75Q7.08 1.75 7.328 1.77 7.803 1.809 8.127 1.974 8.722 2.278 9.026 2.873 9.191 3.197 9.23 3.672 9.25 3.92 9.25 4.571L9.25 4.75 14.75 4.75 14.75 4.571Q14.75 3.92 14.77 3.672 14.809 3.197 14.974 2.873 15.278 2.278 15.873 1.974 16.197 1.809 16.672 1.77 16.92 1.75 17.571 1.75ZM19.429 2.25 17.571 2.25Q16.941 2.25 16.712 2.269 16.336 2.299 16.1 2.42 15.649 2.649 15.42 3.1 15.299 3.336 15.269 3.712 15.25 3.941 15.25 4.571L15.25 5.25 8.75 5.25 8.75 4.571Q8.75 3.941 8.731 3.712 8.701 3.336 8.58 3.1 8.351 2.649 7.9 2.42 7.664 2.299 7.288 2.269 7.059 2.25 6.429 2.25L4.571 2.25Q3.941 2.25 3.712 2.269 3.336 2.299 3.1 2.42 2.649 2.649 2.42 3.1 2.299 3.336 2.269 3.712 2.25 3.941 2.25 4.571L2.25 6.429Q2.25 7.059 2.269 7.288 2.299 7.664 2.42 7.9 2.649 8.351 3.1 8.58 3.336 8.701 3.712 8.731 3.941 8.75 4.571 8.75L5.25 8.75 5.25 15.25 4.571 15.25Q3.941 15.25 3.712 15.269 3.336 15.299 3.1 15.42 2.649 15.649 2.42 16.1 2.299 16.336 2.269 16.712 2.25 16.941 2.25 17.571L2.25 19.429Q2.25 20.059 2.269 20.288 2.299 20.664 2.42 20.9 2.649 21.351 3.1 21.58 3.336 21.701 3.712 21.731 3.941 21.75 4.571 21.75L6.429 21.75Q7.059 21.75 7.288 21.731 7.664 21.701 7.9 21.58 8.351 21.351 8.58 20.9 8.701 20.664 8.731 20.288 8.75 20.059 8.75 19.429L8.75 18.75 15.25 18.75 15.25 19.429Q15.25 20.06 15.269 20.288 15.299 20.664 15.42 20.9 15.649 21.351 16.1 21.58 16.336 21.701 16.712 21.731 16.941 21.75 17.571 21.75L19.429 21.75Q20.059 21.75 20.288 21.731 20.664 21.701 20.9 21.58 21.351 21.351 21.58 20.9 21.701 20.664 21.731 20.288 21.75 20.059 21.75 19.429L21.75 17.571Q21.75 16.941 21.731 16.712 21.701 16.336 21.58 16.1 21.351 15.649 20.9 15.42 20.664 15.299 20.288 15.269 20.06 15.25 19.429 15.25L18.75 15.25 18.75 8.75 19.429 8.75Q20.059 8.75 20.288 8.731 20.664 8.701 20.9 8.58 21.351 8.351 21.58 7.9 21.701 7.664 21.731 7.288 21.75 7.059 21.75 6.429L21.75 4.571Q21.75 3.941 21.731 3.712 21.701 3.336 21.58 3.1 21.351 2.649 20.9 2.42 20.664 2.299 20.288 2.269 20.059 2.25 19.429 2.25ZM18.25 8.75 18.25 15.25 17.571 15.25Q16.94 15.25 16.712 15.269 16.336 15.299 16.1 15.42 15.649 15.649 15.42 16.1 15.299 16.336 15.269 16.712 15.25 16.94 15.25 17.571L15.25 18.25 8.75 18.25 8.75 17.571Q8.75 16.941 8.731 16.712 8.701 16.336 8.58 16.1 8.351 15.649 7.9 15.42 7.664 15.299 7.288 15.269 7.059 15.25 6.429 15.25L5.75 15.25 5.75 8.75 6.429 8.75Q7.059 8.75 7.288 8.731 7.664 8.701 7.9 8.58 8.351 8.351 8.58 7.9 8.701 7.664 8.731 7.288 8.75 7.059 8.75 6.429L8.75 5.75 15.25 5.75 15.25 6.429Q15.25 7.059 15.269 7.288 15.299 7.664 15.42 7.9 15.649 8.351 16.1 8.58 16.336 8.701 16.712 8.731 16.941 8.75 17.571 8.75ZM17.75 9.25 17.571 9.25Q16.92 9.25 16.672 9.23 16.197 9.191 15.873 9.026 15.278 8.722 14.974 8.127 14.809 7.803 14.77 7.328 14.75 7.08 14.75 6.429L14.75 6.25 9.25 6.25 9.25 6.429Q9.25 7.08 9.23 7.328 9.191 7.803 9.026 8.127 8.722 8.722 8.127 9.026 7.803 9.191 7.328 9.23 7.08 9.25 6.429 9.25L6.25 9.25 6.25 14.75 6.429 14.75Q7.08 14.75 7.328 14.77 7.803 14.809 8.127 14.974 8.722 15.278 9.026 15.873 9.191 16.197 9.23 16.672 9.25 16.92 9.25 17.571L9.25 17.75 14.75 17.75 14.75 17.571Q14.75 16.92 14.77 16.672 14.809 16.197 14.974 15.873 15.278 15.278 15.873 14.974 16.197 14.809 16.672 14.77 16.92 14.75 17.571 14.75L17.75 14.75 17.75 9.25ZM16.327 15.865Q16.475 15.79 16.753 15.767 16.961 15.75 17.591 15.75L19.409 15.75Q20.039 15.75 20.247 15.767 20.525 15.79 20.673 15.865 20.979 16.021 21.135 16.327 21.21 16.475 21.233 16.753 21.25 16.961 21.25 17.591L21.25 19.409Q21.25 20.039 21.233 20.247 21.21 20.525 21.135 20.673 20.979 20.979 20.673 21.135 20.525 21.21 20.247 21.233 20.039 21.25 19.409 21.25L17.591 21.25Q16.961 21.25 16.753 21.233 16.475 21.21 16.327 21.135 16.021 20.979 15.865 20.673 15.79 20.525 15.767 20.247 15.75 20.039 15.75 19.409L15.75 17.591Q15.75 16.961 15.767 16.753 15.79 16.475 15.865 16.327 16.021 16.021 16.327 15.865ZM3.753 15.767Q3.961 15.75 4.591 15.75L6.409 15.75Q7.039 15.75 7.247 15.767 7.525 15.79 7.673 15.865 7.979 16.021 8.135 16.327 8.21 16.475 8.233 16.753 8.25 16.961 8.25 17.591L8.25 19.409Q8.25 20.039 8.233 20.247 8.21 20.525 8.135 20.673 7.979 20.979 7.673 21.135 7.525 21.21 7.247 21.233 7.039 21.25 6.409 21.25L4.591 21.25Q3.961 21.25 3.753 21.233 3.475 21.21 3.327 21.135 3.021 20.979 2.865 20.673 2.79 20.525 2.767 20.247 2.75 20.039 2.75 19.409L2.75 17.591Q2.75 16.961 2.767 16.753 2.79 16.475 2.865 16.327 3.021 16.021 3.327 15.865 3.475 15.79 3.753 15.767ZM16.753 2.767Q16.961 2.75 17.591 2.75L19.409 2.75Q20.039 2.75 20.247 2.767 20.525 2.79 20.673 2.865 20.979 3.021 21.135 3.327 21.21 3.475 21.233 3.753 21.25 3.961 21.25 4.591L21.25 6.409Q21.25 7.039 21.233 7.247 21.21 7.525 21.135 7.673 20.979 7.979 20.673 8.135 20.525 8.21 20.247 8.233 20.039 8.25 19.409 8.25L17.591 8.25Q16.961 8.25 16.753 8.233 16.475 8.21 16.327 8.135 16.021 7.979 15.865 7.673 15.79 7.525 15.767 7.247 15.75 7.039 15.75 6.409L15.75 4.591Q15.75 3.961 15.767 3.753 15.79 3.475 15.865 3.327 16.021 3.021 16.327 2.865L16.327 2.865Q16.475 2.79 16.753 2.767ZM4.591 2.75 6.409 2.75Q7.039 2.75 7.247 2.767 7.525 2.79 7.673 2.865 7.979 3.021 8.135 3.327 8.21 3.475 8.233 3.753 8.25 3.961 8.25 4.591L8.25 6.409Q8.25 7.039 8.233 7.247 8.21 7.525 8.135 7.673 7.979 7.979 7.673 8.135 7.525 8.21 7.247 8.233 7.039 8.25 6.409 8.25L4.591 8.25Q3.961 8.25 3.753 8.233 3.475 8.21 3.327 8.135 3.021 7.979 2.865 7.673 2.79 7.525 2.767 7.247 2.75 7.039 2.75 6.409L2.75 4.591Q2.75 3.961 2.767 3.753 2.79 3.475 2.865 3.327 3.021 3.021 3.327 2.865 3.475 2.79 3.753 2.767L3.753 2.767Q3.961 2.75 4.591 2.75ZM16.554 16.311Q16.393 16.393 16.311 16.554 16.28 16.614 16.265 16.794 16.25 16.981 16.25 17.591L16.25 19.409Q16.25 20.019 16.265 20.206 16.28 20.386 16.311 20.446 16.393 20.607 16.554 20.689 16.614 20.72 16.794 20.735 16.981 20.75 17.591 20.75L19.409 20.75Q20.019 20.75 20.206 20.735 20.386 20.72 20.446 20.689 20.607 20.607 20.689 20.446 20.72 20.386 20.735 20.206 20.75 20.019 20.75 19.409L20.75 17.591Q20.75 16.981 20.735 16.794 20.72 16.614 20.689 16.554 20.607 16.393 20.446 16.311 20.386 16.28 20.206 16.265 20.019 16.25 19.409 16.25L17.591 16.25Q16.981 16.25 16.794 16.265 16.614 16.28 16.554 16.311ZM16.554 3.311Q16.393 3.393 16.311 3.554 16.28 3.614 16.265 3.794 16.25 3.981 16.25 4.591L16.25 6.409Q16.25 7.019 16.265 7.206 16.28 7.386 16.311 7.446 16.393 7.607 16.554 7.689 16.614 7.72 16.794 7.735 16.981 7.75 17.591 7.75L19.409 7.75Q20.019 7.75 20.206 7.735 20.386 7.72 20.446 7.689 20.607 7.607 20.689 7.446 20.72 7.386 20.735 7.206 20.75 7.019 20.75 6.409L20.75 4.591Q20.75 3.981 20.735 3.794 20.72 3.614 20.689 3.554 20.607 3.393 20.446 3.311 20.386 3.28 20.206 3.265 20.019 3.25 19.409 3.25L17.591 3.25Q16.981 3.25 16.794 3.265 16.614 3.28 16.554 3.311ZM3.794 16.265Q3.614 16.28 3.554 16.311 3.393 16.393 3.311 16.554 3.28 16.614 3.265 16.794 3.25 16.981 3.25 17.591L3.25 19.409Q3.25 20.019 3.265 20.206 3.28 20.386 3.311 20.446 3.393 20.607 3.554 20.689 3.614 20.72 3.794 20.735 3.981 20.75 4.591 20.75L6.409 20.75Q7.019 20.75 7.206 20.735 7.386 20.72 7.446 20.689 7.607 20.607 7.689 20.446 7.72 20.386 7.735 20.206 7.75 20.019 7.75 19.409L7.75 17.591Q7.75 16.981 7.735 16.794 7.72 16.614 7.689 16.554 7.607 16.393 7.446 16.311 7.386 16.28 7.206 16.265 7.019 16.25 6.409 16.25L4.591 16.25Q3.981 16.25 3.794 16.265ZM3.794 3.265Q3.614 3.28 3.554 3.311 3.393 3.393 3.311 3.554 3.28 3.614 3.265 3.794 3.25 3.981 3.25 4.591L3.25 6.409Q3.25 7.019 3.265 7.206 3.28 7.386 3.311 7.446 3.393 7.607 3.554 7.689 3.614 7.72 3.794 7.735 3.981 7.75 4.591 7.75L6.409 7.75Q7.019 7.75 7.206 7.735 7.386 7.72 7.446 7.689 7.607 7.607 7.689 7.446 7.72 7.386 7.735 7.206 7.75 7.019 7.75 6.409L7.75 4.591Q7.75 3.981 7.735 3.794 7.72 3.614 7.689 3.554 7.607 3.393 7.446 3.311 7.386 3.28 7.206 3.265 7.019 3.25 6.409 3.25L4.591 3.25Q3.981 3.25 3.794 3.265Z";
-export const iconDiscord = "M18.942,6.256 C17.647,5.669 16.263,5.243 14.816,5 C14.638,5.311 14.43,5.73 14.287,6.063 C12.749,5.839 11.225,5.839 9.714,6.063 C9.571,5.73 9.359,5.311 9.179,5 C7.731,5.243 6.345,5.671 5.05,6.259 C2.439,10.082 1.731,13.811 2.085,17.486 C3.817,18.739 5.495,19.5 7.145,19.998 C7.553,19.455 7.916,18.878 8.229,18.269 C7.633,18.049 7.062,17.779 6.523,17.464 C6.666,17.361 6.806,17.254 6.941,17.143 C10.232,18.635 13.807,18.635 17.059,17.143 C17.196,17.254 17.336,17.361 17.477,17.464 C16.936,17.78 16.364,18.051 15.768,18.271 C16.081,18.878 16.442,19.457 16.851,20 C18.503,19.502 20.183,18.741 21.915,17.486 C22.33,13.225 21.206,9.531 18.942,6.256 Z M8.678,15.226 C7.69,15.226 6.88,14.332 6.88,13.244 C6.88,12.156 7.672,11.261 8.678,11.261 C9.683,11.261 10.493,12.154 10.476,13.244 C10.477,14.332 9.683,15.226 8.678,15.226 Z M15.322,15.226 C14.334,15.226 13.524,14.332 13.524,13.244 C13.524,12.156 14.317,11.261 15.322,11.261 C16.328,11.261 17.138,12.154 17.12,13.244 C17.12,14.332 16.328,15.226 15.322,15.226 Z";
+export const iconDiscord = "M14.554,4.449 L14.899,4.507 C16.389,4.757 17.815,5.196 19.148,5.801 L19.274,5.858 L19.353,5.972 C21.678,9.336 22.84,13.146 22.413,17.535 L22.391,17.759 L22.208,17.891 C20.427,19.181 18.696,19.966 16.996,20.479 L16.662,20.58 L16.452,20.301 C16.025,19.734 15.649,19.131 15.323,18.5 L15.237,18.333 C13.109,18.907 10.902,18.905 8.761,18.328 L8.674,18.498 C8.348,19.132 7.969,19.733 7.545,20.299 L7.336,20.578 L7.001,20.477 C5.303,19.965 3.573,19.18 1.792,17.891 L1.609,17.759 L1.587,17.534 C1.223,13.75 1.956,9.903 4.637,5.977 L4.716,5.862 L4.843,5.804 C6.177,5.198 7.605,4.757 9.097,4.507 L9.439,4.45 L9.613,4.751 C9.742,4.975 9.887,5.252 10.015,5.518 C11.329,5.354 12.653,5.354 13.985,5.518 C14.11,5.254 14.253,4.978 14.382,4.752 Z M5.384,6.657 C2.961,10.258 2.272,13.757 2.562,17.212 C4.062,18.269 5.518,18.949 6.952,19.414 C7.155,19.127 7.344,18.83 7.52,18.525 C7.089,18.34 6.671,18.13 6.271,17.896 L5.604,17.508 L6.231,17.058 C6.366,16.962 6.497,16.86 6.625,16.756 L6.865,16.56 L7.147,16.688 C10.308,18.12 13.732,18.119 16.851,16.689 L17.132,16.56 L17.373,16.755 C17.503,16.86 17.637,16.962 17.771,17.06 L18.391,17.51 L17.729,17.896 C17.327,18.131 16.909,18.342 16.476,18.527 C16.652,18.832 16.842,19.129 17.044,19.416 C18.48,18.951 19.938,18.271 21.438,17.212 C21.776,13.224 20.719,9.755 18.609,6.655 C17.495,6.162 16.314,5.79 15.083,5.556 C14.961,5.79 14.839,6.045 14.747,6.261 L14.595,6.613 L14.215,6.558 C12.724,6.341 11.249,6.341 9.788,6.558 L9.407,6.614 L9.255,6.261 C9.163,6.046 9.038,5.79 8.914,5.555 C7.682,5.79 6.499,6.163 5.384,6.657 Z M6.38,13.244 C6.38,11.931 7.348,10.761 8.678,10.761 C9.999,10.761 10.995,11.92 10.976,13.248 C10.975,14.561 10.005,15.726 8.678,15.726 C7.369,15.726 6.38,14.561 6.38,13.244 Z M13.024,13.244 C13.024,11.931 13.993,10.761 15.322,10.761 C16.643,10.761 17.639,11.921 17.62,13.248 C17.618,14.561 16.65,15.726 15.322,15.726 C14.014,15.726 13.024,14.561 13.024,13.244 Z M8.678,11.761 C7.997,11.761 7.38,12.381 7.38,13.244 C7.38,14.103 8.011,14.726 8.678,14.726 C9.36,14.726 9.977,14.106 9.976,13.245 L9.976,13.236 C9.989,12.386 9.365,11.761 8.678,11.761 Z M15.322,11.761 C14.641,11.761 14.024,12.381 14.024,13.244 C14.024,14.103 14.655,14.726 15.322,14.726 C16.004,14.726 16.62,14.106 16.62,13.244 L16.62,13.236 C16.634,12.386 16.01,11.761 15.322,11.761 Z";
 export const iconDollar = "M12.5,2 L12.5,4.013 C13.233,4.059 13.893,4.235 14.479,4.542 C15.168,4.898 15.717,5.391 16.122,6.02 C16.528,6.644 16.753,7.358 16.797,8.157 L16.803,8.265 L15.719,8.265 L15.711,8.172 C15.636,7.245 15.271,6.5 14.617,5.93 C14.058,5.439 13.354,5.157 12.5,5.088 L12.5,11.405 L13.052,11.573 C13.61,11.749 14.127,11.953 14.599,12.185 C15.079,12.418 15.497,12.697 15.852,13.023 C16.214,13.346 16.495,13.726 16.694,14.163 C16.899,14.602 17,15.106 17,15.672 C17,16.499 16.792,17.241 16.376,17.897 C15.965,18.553 15.379,19.068 14.622,19.444 C14.003,19.749 13.295,19.928 12.5,19.982 L12.5,21 L11.5,21 L11.5,19.986 C10.726,19.94 10.036,19.775 9.431,19.49 C8.712,19.149 8.141,18.674 7.72,18.065 C7.299,17.451 7.062,16.744 7.007,15.949 L7,15.84 L8.123,15.84 L8.13,15.934 C8.177,16.549 8.366,17.078 8.696,17.525 C9.031,17.968 9.478,18.312 10.038,18.555 C10.474,18.745 10.962,18.861 11.5,18.903 L11.5,12.249 L11.013,12.097 C9.846,11.731 8.944,11.228 8.317,10.584 C7.69,9.935 7.376,9.126 7.376,8.163 C7.376,7.356 7.584,6.638 8.001,6.011 C8.417,5.382 8.98,4.89 9.689,4.534 C10.24,4.259 10.844,4.09 11.5,4.028 L11.5,2 Z M12.5,12.561 L12.5,18.899 C13.067,18.85 13.578,18.72 14.032,18.511 C14.626,18.234 15.086,17.851 15.417,17.364 C15.747,16.879 15.913,16.314 15.913,15.665 C15.913,15.143 15.794,14.702 15.561,14.337 C15.325,13.97 14.974,13.649 14.502,13.375 C14.033,13.101 13.446,12.854 12.742,12.636 Z M11.5,5.103 C11.025,5.157 10.592,5.281 10.201,5.473 C9.655,5.74 9.23,6.103 8.923,6.561 C8.617,7.016 8.463,7.539 8.463,8.133 C8.463,8.514 8.533,8.855 8.67,9.159 C8.814,9.46 9.015,9.731 9.276,9.972 C9.539,10.209 9.851,10.418 10.216,10.599 C10.582,10.781 10.987,10.941 11.431,11.08 L11.5,11.101 Z";
 export const iconDotsVertical = "M14,19 C14,20.105 13.105,21 12,21 C10.895,21 10,20.105 10,19 C10,17.895 10.895,17 12,17 C13.105,17 14,17.895 14,19 Z M12,14 C10.895,14 10,13.105 10,12 C10,10.895 10.895,10 12,10 C13.105,10 14,10.895 14,12 C14,13.105 13.105,14 12,14 Z M12,7 C10.895,7 10,6.105 10,5 C10,3.895 10.895,3 12,3 C13.105,3 14,3.895 14,5 C14,6.105 13.105,7 12,7 Z M12,13 C12.552,13 13,12.552 13,12 C13,11.448 12.552,11 12,11 C11.448,11 11,11.448 11,12 C11,12.552 11.448,13 12,13 Z M13,19 C13,18.448 12.552,18 12,18 C11.448,18 11,18.448 11,19 C11,19.552 11.448,20 12,20 C12.552,20 13,19.552 13,19 Z M12,6 C12.552,6 13,5.552 13,5 C13,4.448 12.552,4 12,4 C11.448,4 11,4.448 11,5 C11,5.552 11.448,6 12,6 Z";
 export const iconDownload = "M11.553 18.056C11.358 17.996 11.199 17.891 11.051 17.771 10.911 17.659 10.753 17.51 10.572 17.339L5.5 12.549C5.086 12.158 4.752 11.843 4.518 11.581 4.292 11.33 4.077 11.044 4.03 10.71 3.956 10.184 4.165 9.658 4.58 9.327 4.844 9.117 5.195 9.057 5.532 9.029 5.882 9 6.341 9 6.911 9L6.935 9C7.206 9 7.382 9 7.515 8.989 7.643 8.98 7.691 8.963 7.715 8.952 7.819 8.902 7.902 8.818 7.952 8.715 7.963 8.691 7.98 8.643 7.989 8.515 8 8.382 8 8.206 8 7.935L8 4.081C8 3.817 8 3.59 8.015 3.402 8.031 3.205 8.067 3.008 8.163 2.819 8.307 2.537 8.537 2.307 8.819 2.163 9.008 2.067 9.205 2.031 9.402 2.015 9.59 2 9.817 2 10.081 2L13.919 2C14.183 2 14.41 2 14.597 2.015 14.795 2.031 14.992 2.067 15.181 2.163 15.463 2.307 15.693 2.537 15.837 2.819 15.933 3.008 15.969 3.205 15.985 3.402 16 3.59 16 3.817 16 4.081L16 7.935C16 8.206 16 8.382 16.011 8.515 16.021 8.643 16.037 8.691 16.048 8.715 16.098 8.818 16.181 8.902 16.285 8.952 16.309 8.963 16.357 8.98 16.485 8.989 16.618 9 16.794 9 17.065 9L17.089 9C17.659 9 18.118 9 18.468 9.029 18.805 9.057 19.156 9.117 19.42 9.327 19.835 9.658 20.044 10.184 19.97 10.71 19.923 11.044 19.708 11.33 19.482 11.581 19.248 11.843 18.914 12.158 18.5 12.549L13.428 17.339C13.247 17.51 13.089 17.659 12.949 17.771 12.801 17.891 12.642 17.996 12.447 18.056 12.156 18.147 11.844 18.147 11.553 18.056ZM12.149 17.102C12.175 17.094 12.222 17.073 12.323 16.992 12.428 16.908 12.556 16.787 12.755 16.599L17.796 11.839C18.231 11.427 18.535 11.14 18.738 10.913 18.838 10.802 18.901 10.719 18.939 10.656 18.977 10.595 18.98 10.57 18.98 10.57 19.005 10.395 18.935 10.219 18.796 10.109 18.796 10.109 18.777 10.093 18.708 10.074 18.637 10.055 18.535 10.038 18.385 10.026 18.081 10 17.664 10 17.065 10L17.046 10C16.799 10 16.585 10 16.408 9.986 16.222 9.972 16.037 9.941 15.856 9.855 15.545 9.707 15.293 9.456 15.145 9.144 15.059 8.963 15.028 8.778 15.014 8.592 15 8.415 15 8.201 15 7.953L15 4.1C15 3.812 15 3.626 14.988 3.484 14.977 3.348 14.958 3.298 14.946 3.273 14.898 3.179 14.821 3.102 14.727 3.055 14.702 3.042 14.652 3.023 14.516 3.012 14.374 3 14.188 3 13.9 3L10.1 3C9.812 3 9.626 3 9.484 3.012 9.348 3.023 9.298 3.042 9.273 3.055 9.179 3.102 9.102 3.179 9.054 3.273 9.042 3.298 9.023 3.348 9.012 3.484 9 3.626 9 3.812 9 4.1L9 7.954C9 8.201 9 8.415 8.986 8.592 8.972 8.778 8.941 8.963 8.855 9.144 8.707 9.456 8.456 9.707 8.144 9.855 7.963 9.941 7.778 9.972 7.592 9.986 7.415 10 7.201 10 6.953 10L6.935 10C6.336 10 5.918 10 5.615 10.026 5.465 10.038 5.363 10.055 5.292 10.074 5.223 10.093 5.203 10.109 5.203 10.109 5.065 10.219 4.995 10.395 5.02 10.57 5.02 10.57 5.023 10.595 5.061 10.656 5.099 10.719 5.162 10.802 5.262 10.913 5.465 11.14 5.769 11.427 6.204 11.839L11.245 16.599C11.444 16.787 11.572 16.908 11.677 16.992 11.778 17.073 11.825 17.094 11.851 17.102 11.948 17.132 12.052 17.132 12.149 17.102ZM21 21 21 22 3 22 3 21 21 21Z";
@@ -117,6 +118,7 @@ export const iconTextBoxMinus = "M7.478,3 L16.522,3 C17.204,3 17.746,3 18.183,3.
 export const iconTextBoxPlus = "M7.478,3 L16.522,3 C17.204,3 17.746,3 18.183,3.036 C18.631,3.072 19.012,3.149 19.362,3.327 C19.927,3.615 20.385,4.074 20.673,4.638 C20.851,4.988 20.928,5.369 20.964,5.817 C21,6.255 21,6.796 21,7.478 L21,12 L20,12 L20,7.5 C20,6.792 20,6.291 19.968,5.899 C19.936,5.513 19.876,5.277 19.782,5.092 C19.59,4.716 19.284,4.41 18.908,4.218 C18.723,4.124 18.487,4.064 18.101,4.032 C17.709,4 17.208,4 16.5,4 L7.5,4 C6.792,4 6.291,4 5.899,4.032 C5.513,4.064 5.277,4.124 5.092,4.218 C4.716,4.41 4.41,4.716 4.218,5.092 C4.124,5.277 4.064,5.513 4.032,5.899 C4,6.291 4,6.792 4,7.5 L4,16.5 C4,17.208 4,17.709 4.032,18.101 C4.064,18.487 4.124,18.723 4.218,18.908 C4.41,19.284 4.716,19.59 5.092,19.782 C5.277,19.876 5.513,19.936 5.899,19.968 C6.291,20 6.792,20 7.5,20 L12,20 L12,21 L7.478,21 C6.796,21 6.255,21 5.817,20.964 C5.369,20.928 4.988,20.851 4.638,20.673 C4.074,20.385 3.615,19.927 3.327,19.362 C3.149,19.012 3.072,18.631 3.036,18.183 C3,17.746 3,17.204 3,16.522 L3,7.478 C3,6.796 3,6.254 3.036,5.817 C3.072,5.369 3.149,4.988 3.327,4.638 C3.615,4.074 4.074,3.615 4.638,3.327 C4.988,3.149 5.369,3.072 5.817,3.036 C6.254,3 6.796,3 7.478,3 L7.478,3 Z M17.5,18.5 L14,18.5 L14,17.5 L17.5,17.5 L17.5,14 L18.5,14 L18.5,17.5 L22,17.5 L22,18.5 L18.5,18.5 L18.5,22 L17.5,22 Z M7,8 L7,7 L17,7 L17,8 Z M17,12 L7,12 L7,11 L17,11 Z M7,16 L7,15 L12,15 L12,16 Z";
 export const iconTransfer = "M7.537 1.487C7.838 1.39 8.163 1.39 8.464 1.487 8.666 1.553 8.83 1.667 8.981 1.795 9.124 1.917 9.285 2.078 9.472 2.264L12.501 5.293C12.699 5.491 12.862 5.655 12.983 5.789 13.098 5.918 13.222 6.07 13.287 6.247 13.525 6.885 13.229 7.598 12.61 7.882 12.439 7.96 12.243 7.981 12.071 7.99 11.891 8 11.66 8 11.38 8L11.365 8C10.99 8 10.911 8.005 10.859 8.02 10.696 8.068 10.569 8.196 10.521 8.359 10.505 8.41 10.5 8.49 10.5 8.864L10.5 12.424C10.5 12.686 10.5 12.911 10.485 13.097 10.469 13.295 10.433 13.492 10.337 13.681 10.193 13.963 9.963 14.193 9.681 14.336 9.492 14.433 9.295 14.468 9.098 14.485 8.91 14.5 8.683 14.5 8.42 14.5L7.581 14.5C7.317 14.5 7.09 14.5 6.903 14.485 6.705 14.468 6.508 14.433 6.319 14.336 6.037 14.193 5.808 13.963 5.664 13.681 5.567 13.492 5.532 13.295 5.516 13.097 5.5 12.91 5.5 12.683 5.5 12.419L5.5 8.864C5.5 8.49 5.495 8.41 5.48 8.359 5.432 8.196 5.304 8.068 5.141 8.02 5.09 8.005 5.01 8 4.636 8L4.621 8C4.341 8 4.11 8 3.93 7.99 3.757 7.981 3.562 7.96 3.39 7.882 2.771 7.598 2.476 6.885 2.713 6.247 2.779 6.07 2.903 5.918 3.018 5.789 3.138 5.655 3.302 5.491 3.5 5.293L6.529 2.264C6.715 2.078 6.876 1.917 7.019 1.795 7.171 1.667 7.335 1.553 7.537 1.487ZM8.155 2.439C8.054 2.406 7.946 2.406 7.846 2.439 7.819 2.447 7.77 2.469 7.667 2.558 7.558 2.65 7.426 2.781 7.222 2.985L4.217 5.99C4.006 6.201 3.862 6.345 3.763 6.456 3.67 6.56 3.649 6.598 3.649 6.598L3.65 6.595C3.596 6.743 3.664 6.907 3.807 6.973L3.804 6.971C3.804 6.971 3.846 6.984 3.984 6.992 4.134 7 4.337 7 4.636 7 4.656 7 4.677 7 4.697 7L4.697 7C4.975 6.999 5.214 6.999 5.424 7.061 5.913 7.205 6.295 7.587 6.439 8.076 6.501 8.286 6.501 8.526 6.5 8.803L6.5 8.864 6.5 12.4C6.5 12.688 6.501 12.874 6.512 13.016 6.523 13.152 6.542 13.202 6.555 13.227 6.603 13.321 6.679 13.397 6.773 13.445 6.798 13.458 6.848 13.477 6.984 13.488 7.126 13.5 7.312 13.5 7.6 13.5L8.4 13.5C8.689 13.5 8.875 13.5 9.016 13.488 9.152 13.477 9.203 13.458 9.227 13.445 9.321 13.397 9.398 13.321 9.446 13.227 9.458 13.202 9.477 13.152 9.488 13.016 9.5 12.874 9.5 12.688 9.5 12.4L9.5 8.864C9.5 8.844 9.5 8.823 9.5 8.803L9.5 8.803C9.5 8.526 9.5 8.286 9.561 8.076 9.705 7.587 10.088 7.205 10.577 7.061 10.786 6.999 11.026 6.999 11.304 7L11.304 7C11.324 7 11.344 7 11.365 7 11.664 7 11.867 7 12.016 6.992 12.155 6.984 12.196 6.971 12.196 6.971L12.194 6.973C12.337 6.907 12.405 6.743 12.35 6.595L12.351 6.598C12.351 6.598 12.331 6.56 12.238 6.456 12.138 6.345 11.994 6.201 11.783 5.99L8.778 2.985C8.574 2.781 8.442 2.65 8.334 2.558 8.23 2.469 8.181 2.447 8.155 2.439ZM16.464 22.512C16.163 22.61 15.838 22.61 15.537 22.512 15.335 22.447 15.17 22.333 15.019 22.204 14.876 22.083 14.715 21.922 14.529 21.736L11.5 18.706C11.302 18.508 11.138 18.345 11.018 18.211 10.903 18.082 10.779 17.93 10.713 17.753 10.476 17.115 10.771 16.401 11.39 16.118 11.562 16.039 11.757 16.019 11.93 16.01 12.11 16 12.341 16 12.621 16L12.636 16C13.01 16 13.09 15.995 13.141 15.979 13.304 15.932 13.432 15.804 13.48 15.641 13.495 15.589 13.5 15.51 13.5 15.135L13.5 11.581C13.5 11.317 13.5 11.09 13.516 10.902 13.532 10.705 13.567 10.508 13.664 10.319 13.807 10.037 14.037 9.807 14.319 9.663 14.508 9.567 14.705 9.531 14.903 9.515 15.09 9.5 15.317 9.5 15.581 9.5L16.42 9.5C16.683 9.5 16.911 9.5 17.098 9.515 17.295 9.531 17.492 9.567 17.681 9.663 17.963 9.807 18.193 10.037 18.337 10.319 18.433 10.508 18.469 10.705 18.485 10.902 18.5 11.088 18.5 11.314 18.5 11.576L18.5 15.135C18.5 15.51 18.505 15.589 18.521 15.641 18.569 15.804 18.696 15.932 18.859 15.979 18.911 15.995 18.99 16 19.365 16L19.38 16C19.66 16 19.891 16 20.071 16.01 20.243 16.019 20.439 16.039 20.61 16.118 21.229 16.401 21.525 17.115 21.287 17.753 21.222 17.93 21.098 18.082 20.983 18.211 20.862 18.345 20.699 18.508 20.501 18.706L17.472 21.736C17.285 21.922 17.124 22.083 16.981 22.204 16.83 22.333 16.666 22.447 16.464 22.512ZM15.846 21.561C15.946 21.594 16.054 21.594 16.155 21.561 16.181 21.553 16.23 21.53 16.334 21.442 16.442 21.35 16.574 21.219 16.778 21.015L19.783 18.01C19.994 17.799 20.138 17.655 20.238 17.543 20.331 17.44 20.351 17.402 20.351 17.402 20.404 17.255 20.336 17.092 20.194 17.027L20.197 17.028C20.197 17.028 20.155 17.016 20.016 17.008 19.867 17 19.664 17 19.365 17L19.303 17C19.026 17 18.786 17 18.577 16.939 18.088 16.795 17.705 16.412 17.561 15.924 17.5 15.714 17.5 15.474 17.5 15.197L17.5 11.6C17.5 11.312 17.5 11.125 17.488 10.984 17.477 10.848 17.458 10.798 17.446 10.773 17.398 10.679 17.321 10.602 17.227 10.554 17.202 10.542 17.152 10.523 17.016 10.512 16.875 10.5 16.688 10.5 16.4 10.5L15.6 10.5C15.312 10.5 15.126 10.5 14.984 10.512 14.848 10.523 14.798 10.542 14.773 10.554 14.679 10.602 14.603 10.679 14.555 10.773 14.542 10.798 14.523 10.848 14.512 10.984 14.501 11.125 14.5 11.312 14.5 11.6L14.5 15.135C14.5 15.156 14.5 15.176 14.5 15.196L14.5 15.197C14.501 15.474 14.501 15.714 14.439 15.924 14.295 16.412 13.913 16.795 13.424 16.939 13.214 17 12.975 17 12.697 17L12.636 17C12.337 17 12.134 17 11.984 17.008 11.845 17.016 11.804 17.028 11.804 17.028L11.807 17.027C11.665 17.092 11.597 17.255 11.649 17.402 11.649 17.402 11.67 17.44 11.763 17.543 11.862 17.655 12.006 17.799 12.217 18.01L15.222 21.015C15.426 21.219 15.558 21.35 15.667 21.442 15.77 21.53 15.819 21.553 15.846 21.561Z";
 export const iconTune = "M15.5 10C13.719 10 12.249 8.67 12.028 6.949L2 6.949 2 6.051 12.028 6.051C12.249 4.33 13.719 3 15.5 3 17.281 3 18.751 4.33 18.972 6.051L22 6.051 22 6.949 18.972 6.949C18.751 8.67 17.281 10 15.5 10ZM18 6.5C18 5.119 16.881 4 15.5 4 14.119 4 13 5.119 13 6.5 13 7.881 14.119 9 15.5 9 16.881 9 18 7.881 18 6.5ZM11.972 17.051 22 17.051 22 17.949 11.972 17.949C11.751 19.67 10.281 21 8.5 21 6.719 21 5.249 19.67 5.028 17.949L2 17.949 2 17.051 5.029 17.051C5.249 15.33 6.719 14 8.5 14 10.281 14 11.751 15.33 11.972 17.051ZM6 17.5C6 18.881 7.119 20 8.5 20 9.881 20 11 18.881 11 17.5 11 16.119 9.881 15 8.5 15 7.119 15 6 16.119 6 17.5Z";
+export const iconUniversity = "M11.826,4.531 C11.938,4.49 12.062,4.49 12.174,4.531 L23.174,8.605 C23.362,8.675 23.49,8.85 23.5,9.051 C23.509,9.251 23.398,9.438 23.217,9.525 L18.611,11.742 L18.611,16.407 C18.611,16.893 18.322,17.311 17.968,17.633 C17.604,17.965 17.103,18.262 16.518,18.511 C15.344,19.011 13.749,19.352 12.002,19.352 C10.255,19.352 8.66,19.011 7.485,18.512 C6.9,18.264 6.398,17.966 6.034,17.635 C5.68,17.313 5.39,16.895 5.389,16.41 L5.389,11.742 L2.722,10.458 L2.722,12.195 C2.749,12.241 2.781,12.298 2.818,12.364 C2.919,12.545 3.053,12.799 3.187,13.094 C3.447,13.667 3.741,14.462 3.741,15.185 C3.741,15.7 3.578,16.134 3.287,16.444 C2.999,16.751 2.611,16.907 2.222,16.907 C1.833,16.907 1.445,16.751 1.157,16.444 C0.867,16.134 0.704,15.7 0.704,15.185 C0.704,14.462 0.998,13.667 1.258,13.094 C1.392,12.799 1.526,12.545 1.626,12.364 C1.663,12.298 1.696,12.241 1.722,12.195 L1.722,9.977 L0.783,9.525 C0.602,9.438 0.491,9.251 0.501,9.051 C0.51,8.85 0.638,8.675 0.826,8.605 Z M2.277,9.134 L12,13.815 L21.723,9.134 L12,5.533 Z M6.389,12.224 L6.389,16.408 C6.389,16.485 6.442,16.654 6.707,16.895 C6.962,17.127 7.356,17.371 7.876,17.592 C8.915,18.033 10.375,18.352 12.002,18.352 C13.629,18.352 15.089,18.033 16.126,17.591 C16.646,17.37 17.04,17.126 17.294,16.894 C17.559,16.653 17.611,16.484 17.611,16.407 L17.611,12.224 L12.217,14.821 C12.08,14.887 11.92,14.887 11.783,14.821 Z M2.222,13.391 C2.204,13.429 2.186,13.468 2.168,13.508 C1.919,14.056 1.704,14.687 1.704,15.185 C1.704,15.486 1.795,15.663 1.887,15.76 C1.98,15.86 2.102,15.907 2.222,15.907 C2.343,15.907 2.464,15.86 2.558,15.76 C2.649,15.663 2.741,15.486 2.741,15.185 C2.741,14.687 2.525,14.056 2.276,13.508 C2.258,13.468 2.24,13.429 2.222,13.391 Z";
 export const iconUpload = "M11.553 2C11.844 1.909 12.156 1.909 12.447 2 12.642 2.061 12.801 2.166 12.949 2.285 13.089 2.397 13.247 2.546 13.428 2.718L18.5 7.507C18.914 7.898 19.248 8.214 19.482 8.475 19.708 8.727 19.923 9.012 19.97 9.346 20.044 9.872 19.835 10.398 19.42 10.729 19.156 10.939 18.805 10.999 18.468 11.027 18.118 11.056 17.658 11.056 17.089 11.056L17.065 11.056C16.794 11.056 16.618 11.057 16.485 11.067 16.357 11.077 16.309 11.094 16.285 11.105 16.181 11.154 16.098 11.238 16.048 11.342 16.037 11.365 16.021 11.413 16.011 11.541 16 11.675 16 15.294 16 15.565L16 15.976C16 16.239 16 16.466 15.985 16.654 15.969 16.851 15.933 17.048 15.837 17.237 15.693 17.52 15.463 17.749 15.181 17.893 14.992 17.989 14.795 18.025 14.597 18.041 14.41 18.056 14.183 18.056 13.919 18.056L10.081 18.056C9.817 18.056 9.59 18.056 9.402 18.041 9.205 18.025 9.008 17.989 8.819 17.893 8.537 17.749 8.307 17.52 8.163 17.237 8.067 17.048 8.031 16.851 8.015 16.654 8 16.467 8 16.239 8 15.976L8 15.565C8 15.294 8 11.675 7.989 11.541 7.98 11.413 7.963 11.365 7.952 11.342 7.902 11.238 7.819 11.154 7.715 11.105 7.691 11.094 7.643 11.077 7.515 11.067 7.382 11.057 7.206 11.056 6.935 11.056L6.911 11.056C6.342 11.056 5.882 11.056 5.532 11.027 5.195 10.999 4.844 10.939 4.58 10.729 4.165 10.398 3.956 9.872 4.03 9.346 4.077 9.012 4.292 8.727 4.518 8.475 4.752 8.214 5.086 7.898 5.5 7.507L10.572 2.718C10.753 2.546 10.911 2.397 11.051 2.285 11.199 2.166 11.358 2.061 11.553 2ZM12.149 2.955C12.052 2.924 11.948 2.924 11.851 2.955 11.825 2.963 11.778 2.983 11.677 3.064 11.572 3.149 11.444 3.269 11.245 3.457L6.204 8.218C5.769 8.629 5.465 8.916 5.262 9.143 5.162 9.255 5.099 9.337 5.061 9.4 5.023 9.461 5.02 9.486 5.02 9.486 4.995 9.661 5.065 9.837 5.203 9.947 5.203 9.947 5.223 9.963 5.292 9.982 5.363 10.001 5.465 10.018 5.615 10.031 5.918 10.056 6.336 10.056 6.935 10.056L6.953 10.056C7.201 10.056 7.415 10.056 7.592 10.07 7.778 10.084 7.963 10.116 8.144 10.201 8.456 10.349 8.707 10.601 8.855 10.913 8.941 11.093 8.972 11.278 8.986 11.465 9 11.641 9 15.299 9 15.547L9 15.956C9 16.245 9 16.431 9.012 16.572 9.023 16.708 9.042 16.759 9.054 16.783 9.102 16.877 9.179 16.954 9.273 17.002 9.298 17.014 9.348 17.033 9.484 17.044 9.626 17.056 9.812 17.056 10.1 17.056L13.9 17.056C14.188 17.056 14.374 17.056 14.516 17.044 14.652 17.033 14.702 17.014 14.727 17.002 14.821 16.954 14.898 16.877 14.946 16.783 14.958 16.759 14.977 16.708 14.988 16.572 15 16.431 15 16.245 15 15.956L15 15.547C15 15.299 15 11.641 15.014 11.465 15.028 11.278 15.059 11.093 15.145 10.913 15.293 10.601 15.545 10.349 15.856 10.201 16.037 10.116 16.222 10.084 16.408 10.07 16.585 10.056 16.799 10.056 17.046 10.056L17.065 10.056C17.664 10.056 18.081 10.056 18.385 10.031 18.535 10.018 18.637 10.001 18.708 9.982 18.777 9.963 18.796 9.947 18.796 9.947 18.935 9.837 19.005 9.661 18.98 9.486 18.98 9.486 18.977 9.461 18.939 9.4 18.901 9.337 18.838 9.255 18.738 9.143 18.535 8.916 18.231 8.629 17.796 8.218L12.755 3.457C12.556 3.269 12.428 3.149 12.323 3.064 12.222 2.983 12.175 2.963 12.149 2.955ZM21 22 3 22 3 21 21 21 21 22Z";
 export const iconVariable = "M7.276,3.053 L7.724,3.947 C6.505,4.557 5.976,4.825 5.563,5.181 C4.862,5.785 4.364,6.59 4.137,7.487 C4.004,8.016 4,8.609 4,9.972 L4,13.028 C4,14.391 4.004,14.984 4.137,15.512 C4.364,16.41 4.862,17.215 5.563,17.819 C5.976,18.175 6.505,18.443 7.724,19.053 L7.276,19.947 L7.205,19.911 C6.078,19.348 5.429,19.023 4.91,18.576 C4.053,17.838 3.445,16.854 3.168,15.757 C3,15.094 3,14.368 3,13.108 L3,9.972 L3,9.892 C3,8.632 3,7.906 3.168,7.242 C3.445,6.146 4.053,5.162 4.91,4.423 C5.429,3.976 6.078,3.652 7.205,3.089 Z M16.724,3.053 L16.795,3.089 C17.922,3.652 18.571,3.976 19.09,4.423 C19.947,5.162 20.555,6.146 20.832,7.242 C21,7.906 21,8.632 21,9.892 L21,13.108 C21,14.368 21,15.094 20.832,15.757 C20.555,16.854 19.947,17.838 19.09,18.576 C18.571,19.023 17.922,19.348 16.795,19.911 L16.724,19.947 L16.276,19.053 C17.495,18.443 18.024,18.175 18.437,17.819 C19.138,17.215 19.636,16.41 19.863,15.512 C19.996,14.984 20,14.391 20,13.028 L20,9.972 C20,8.609 19.996,8.016 19.863,7.487 C19.636,6.59 19.138,5.785 18.437,5.181 C18.024,4.825 17.495,4.557 16.276,3.947 Z M11.375,11.5 L8.6,7.8 L9.4,7.2 L12,10.667 L14.6,7.2 L15.4,7.8 L12.625,11.5 L15.4,15.2 L14.6,15.8 L12,12.333 L9.4,15.8 L8.6,15.2 Z";
 export const iconVideo = "M16,15.3 L16,15.6 C16,16.44 16,16.86 15.837,17.181 C15.693,17.463 15.463,17.693 15.181,17.837 C14.86,18 14.44,18 13.6,18 L4.4,18 C3.56,18 3.14,18 2.819,17.837 C2.537,17.693 2.307,17.463 2.163,17.181 C2,16.86 2,16.44 2,15.6 L2,8.4 C2,7.56 2,7.14 2.163,6.819 C2.307,6.537 2.537,6.307 2.819,6.163 C3.14,6 3.56,6 4.4,6 L13.6,6 C14.44,6 14.86,6 15.181,6.163 C15.463,6.307 15.693,6.537 15.837,6.819 C16,7.14 16,7.56 16,8.4 L16,8.7 L18.011,7.487 C18.524,7.177 18.938,6.927 19.279,6.762 C19.622,6.595 19.965,6.475 20.321,6.505 C20.843,6.548 21.323,6.809 21.638,7.226 C21.854,7.513 21.93,7.868 21.965,8.243 C22,8.615 22,9.093 22,9.682 L22,14.318 C22,14.907 22,15.385 21.965,15.757 C21.93,16.132 21.854,16.487 21.638,16.774 C21.323,17.191 20.843,17.452 20.321,17.496 C19.965,17.525 19.622,17.405 19.279,17.238 C18.939,17.073 18.524,16.823 18.011,16.513 Z M4.4,7 C3.963,7 3.696,7.001 3.496,7.017 C3.403,7.025 3.347,7.034 3.312,7.042 C3.296,7.046 3.286,7.049 3.281,7.051 C3.276,7.053 3.273,7.054 3.273,7.055 C3.179,7.102 3.102,7.179 3.055,7.273 C3.054,7.273 3.053,7.276 3.051,7.281 C3.049,7.286 3.046,7.296 3.042,7.312 C3.034,7.347 3.025,7.403 3.017,7.496 C3.001,7.696 3,7.963 3,8.4 L3,15.6 C3,16.037 3.001,16.304 3.017,16.504 C3.025,16.597 3.034,16.653 3.042,16.688 C3.046,16.704 3.049,16.714 3.051,16.719 C3.053,16.724 3.054,16.727 3.055,16.727 C3.102,16.821 3.179,16.898 3.273,16.945 C3.273,16.946 3.276,16.947 3.281,16.949 C3.286,16.951 3.296,16.954 3.312,16.958 C3.347,16.966 3.403,16.975 3.496,16.983 C3.696,16.999 3.963,17 4.4,17 L13.6,17 C14.036,17 14.304,16.999 14.504,16.983 C14.597,16.975 14.653,16.966 14.688,16.958 C14.704,16.954 14.714,16.951 14.719,16.949 C14.724,16.947 14.727,16.946 14.727,16.945 C14.821,16.898 14.898,16.821 14.946,16.727 C14.946,16.727 14.947,16.724 14.949,16.719 C14.951,16.714 14.954,16.704 14.958,16.688 C14.966,16.653 14.975,16.597 14.983,16.504 C14.999,16.304 15,16.037 15,15.6 L15,8.4 C15,7.963 14.999,7.696 14.983,7.496 C14.975,7.403 14.966,7.347 14.958,7.312 C14.954,7.296 14.951,7.286 14.949,7.281 C14.947,7.276 14.946,7.273 14.946,7.273 C14.898,7.179 14.821,7.102 14.727,7.055 C14.727,7.054 14.724,7.053 14.719,7.051 C14.714,7.049 14.704,7.046 14.688,7.042 C14.653,7.034 14.597,7.025 14.504,7.017 C14.304,7.001 14.036,7 13.6,7 Z M16,14.132 L18.506,15.644 C19.046,15.969 19.421,16.196 19.717,16.339 C20.015,16.484 20.156,16.506 20.238,16.499 C20.48,16.479 20.699,16.358 20.84,16.171 C20.884,16.112 20.939,15.988 20.969,15.664 C21,15.344 21,14.913 21,14.292 L21,9.708 C21,9.086 21,8.656 20.969,8.336 C20.939,8.013 20.884,7.888 20.84,7.829 C20.699,7.642 20.48,7.521 20.238,7.501 C20.156,7.494 20.015,7.516 19.717,7.661 C19.421,7.804 19.046,8.031 18.506,8.356 L16,9.868 Z";
